requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
appdirs==1.4.4
asgiref==3.3.4; python_version >= '3.6'
autoflake==1.4
black==21.4b2
cfgv==3.3.0; python_full_version >= '3.6.1'
click==8.0.1; python_version >= '3.6'
coverage==5.5
distlib==0.3.2
django==3.2.3
djangorestframework==3.12.4
drf-nested-routers==0.93.3
factory-boy==3.2.0
faker==8.1.0
filelock==3.0.12
identify==2.2.10; python_full_version >= '3.6.1'
isort==5.8.0
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.8.1
pre-commit==2.12.1
pyflakes==2.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
regex==2021.4.4
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.4.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
